Fixed Hessian, graphical minimum placement
The Hessian is now calculated with the Python package numdifftools.
Getting it iteratively from the minimization can cause the unit matrix
to be returned instead, setting all errors to sqrt(2).

After calculating the contour the function to be minimized is called
once more at the minimum to reset the nexus cache.